Use JobScheduler for OTA checking in the background
Starting Android 8, AOSP now *loves* to prefer tasks queued with JobScheduler because they can be dynamically scheduled according to Doze.

This sets up the app to use JobScheduler to enqueue background tasks, with a compatible syntax that followed the older Utils.scheduleNotification() method.

This isn't fully tested (*minimum* time of recurrence is 15 minutes), but jobs get queued successfully according to the logs.