Use a non-intrusive API to set resolution
This uses an undocumented android.view.IWindowManager API that forces both screen resolution and density, with a syntax that is very similar to the one that used root.

Since these APIs have been blocked starting Android 9, we set setting flags to temporarily allow the app to use them; we disable those after for security. We also store the default value so we can keep them on if user *intentionally* enabled those APIs (i.e. for another app).

API is a derivative of works by @farmerbb and @tytydraco. Please attribute them.