Huge changes to the addon system
This squashed commit comes with the following changes:

- Re-worked addons download database
This removes the dependency on having the downloads be stored in memory, now using a(n admittedly very basic) database using SharedPreferences. This allows us to track downloads on multiple activities, since addons have an "AddonInfo" activity now. This also allows us to tell the app if the download was ever successful or not, even if the user exits the app.

- Re-worked background operations
This moves away from what was common AsyncTask, and now handles operations in the background using Executors for tracking download progress. AsyncTask is deprecated starting Android 11. This makes background operations faster as well, at least from testing.

- Logic in AddonInfoActivity
Downloads now work in AddonInfoActivity, thanks to the changes with how addon downloads are tracked.

- Some layout improvements
Some margin fixes and just changes that seem to make the addon activities look better (in my opinion).