Introduce Renoir
Renoir is an overlay-driven feature inspired by Android 12's Monet color-changing system theme. It detects the primary (or secondary) color of the wallpaper and calculates the nearest color based on L*a*b-based CIECAM2000 with the color palette pulled from Samsung One UI's design guidelines.

It gets triggered when the wallpaper is changed, as well as a theme is applied through the Galaxy Store (disabling the service).

It's unfinished on the system side, but color calculation is pretty much finished.