MaverickService: Introduce MaverickService
MaverickService (an obvious reference to MegaMan lmao) is a service that handles kernel support for 'deny_new_usb'; a feature that allows a device to block new USB-OTG devices, preventing BadUSB attacks.

See https://blog.lizzie.io/preventing-usb-attacks-with-grsecurity.html.

There's currently no UI to configure this but the service is somewhat complete. Also to note that this requires kernel support, hence the check in the service. Fresh Core support to follow.